refactor(ui): 抽 LogConsole，并扩展日志缓冲给应用实例
Bot 日志页与应用端日志 Tab 共用同一控制台；缓冲按实例分流。
